Separated Value tests and add some tests
