+++ test/base_test.rb
@@ -32,6 +32,12 @@ class BaseTest < Minitest::Test
     assert_equal statuses.first, Status.draft
   end
 
+  def test_required_id
+    assert_raises 'Enumeration id is required' do
+      Class.new.value draft: { name: 'Draft' }
+    end
+  end
+
   def test_duplicated_id
     assert_raises 'Duplicate id 1' do
       Class.new.values draft: { id: 1, name: 'Draft' },
@@ -55,47 +61,5 @@ class BaseTest < Minitest::Test
       statuses.map(&:visible)
       statuses.map(&:deleted)
     end
-  end
-
-  def test_with_defined_custom_attributes_visible
-    status = Status.find(:none)
-
-    assert_equal true, status.visible
-  end
-
-  def test_with_defined_custom_attributes_deleted
-    status = Status.find(:deleted)
-
-    assert_equal true, status.deleted
-  end
-
-  def test_without_defined_custom_attributes
-    status = Status.find(:draft)
-
-    assert_equal nil, status.visible
-  end
-
-  def test_equal_by_id
-    status = Status.find(:draft)
-
-    assert_equal true, status == 1
-  end
-
-  def test_equal_by_symbol
-    status = Status.draft
-
-    assert_equal true, status == :draft
-  end
-
-  def test_equal_by_enumeration
-    status = Status.draft
-
-    assert_equal true, status == Status.draft
-  end
-
-  def test_not_equal_by_enumeration
-    status = Status.draft
-
-    assert_equal false, status == Status.published
   end
 end
