+++ test/value_test.rb
@@ -1,0 +1,45 @@
+require_relative 'test_helper'
+
+class ValueTest < Minitest::Test
+  def test_equal_by_id
+    status = Status.find(:draft)
+
+    assert_equal true, status == 1
+  end
+
+  def test_equal_by_symbol
+    status = Status.draft
+
+    assert_equal true, status == :draft
+  end
+
+  def test_equal_by_enumeration
+    status = Status.draft
+
+    assert_equal true, status == Status.draft
+  end
+
+  def test_not_equal_by_enumeration
+    status = Status.draft
+
+    assert_equal false, status == Status.published
+  end
+
+  def test_with_defined_custom_attributes_visible
+    status = Status.find(:none)
+
+    assert_equal true, status.visible
+  end
+
+  def test_with_defined_custom_attributes_deleted
+    status = Status.find(:deleted)
+
+    assert_equal true, status.deleted
+  end
+
+  def test_without_defined_custom_attributes
+    status = Status.find(:draft)
+
+    assert_equal nil, status.visible
+  end
+end
